requirements.txt
uvicorn
python-dotenv
fastapi
pydantic
numpy
